+++ src/Provider/ScheduleProvider.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren, useMemo, useState } from "react";
-import dummyScheduleMap from "./dummyScheduleMap.ts";
-import { useAutoCallback } from "./hooks/useAutoCallback.tsx";
-import { Schedule } from "./types.ts";
+import dummyScheduleMap from "../dummyScheduleMap.ts";
+import { useAutoCallback } from "../hooks/useAutoCallback.tsx";
+import { Schedule } from "../types.ts";
 
 const ScheduleStateContext = React.createContext<Record<
   string,
