+++ src/components/Schedule/ScheduleContainer.tsx
@@ -2,8 +2,8 @@ import { Stack } from "@chakra-ui/react";
 
 import { Dispatch, memo, SetStateAction } from "react";
 import { useAutoCallback } from "../../hooks/useAutoCallback.tsx";
-import { useScheduleActions } from "../../ScheduleContext.tsx";
-import ScheduleDndProvider from "../../ScheduleDndProvider.tsx";
+import { useScheduleActions } from "../../Provider/ScheduleProvider.tsx";
+import ScheduleDndProvider from "../../Provider/ScheduleDndProvider.tsx";
 import ScheduleTable from "../../ScheduleTable.tsx";
 import { Schedule } from "../../types.ts";
 import ScheduleTableHeader from "./ScheduleTableHeader.tsx";
